auto/build: clone more build output to the log file.
Previously we would start cloning the output later in auto/build, because we
needed to do a bunch of things first, until we could figure out what the name of
the build log file was. This is not the case anymore and we can thus generate
a more complete build log.

A future improvement would be to move this to the Rakefile or to build-tails,
but that's more involved and I'm after low-risk, easy incremental improvements
at the moment.