Fix FTBFS with tagged 3.14 snapshots, due to procmail.
procmail is listed with packages said to be “Priority: standard” but
it's only “Priority: optional” in stretch. With monkeysphere going away
(in 2cee79492549867921a81ec4ab7066d73ef54bee), this package is no longer
pull and no longer visible in tagged snapshots, which leads to this
FTBFS:

    E: Unable to locate package procmail
    E: config/chroot_local-hooks/98-remove_unwanted_packages failed (exit non-zero). You should check for errors.

Hotfix for 3.14: switch to mutt-style removal, and trigger a purge for
procmail only if it's (known and) installed.