Temporarily disable building virtualbox-guest-dkms (refs: #16741)
Due to https://bugs.debian.org/929399, the module fails to build with
virtualbox-guest-dkms 6.0.8-dfsg-5, while it built just fine with 6.0.8-dfsg-4.
Let's disable it temporarily, hoping the bug is fixed and we can revert this
commit very soon; if that does not happen, we'll probably need to upload -4 to
our custom APT repository.