AppArmor: allow cups-brf, driverless, and gutenprint53+usb backends (refs: #15030)
Technically, cups-brf and driverless are not third-party and should be confined
more strictly with "ixr", under the cupsd profile. But I don't know how to to
test these backends and confining them more strictly may break them.

Anyway, that's an upstream matter: the purpose of our Tails-specific patch is to
replace the third party backends /usr/lib/cups/backend/* catch all rule, that
doesn't work for us, and not to keep the list of backends which come with CUPS
up-to-date.