Fix "Tor check" button position on home/testing (refs: #12003)
The way this button is implemented implicitly relies on the fact its the first
content in <div id="content" role="main">. If that's not the case, then this
button will be placed somewhere in the middle of the sidebar, instead of on top
of the download button, as intended.

So, let's refactor home.html so common bits are in inlined sub-pages,
in a way that allows me to order the Tor check button properly
on home/testing.