Fix screen locker not working in Buster (refs: #16763)
We used the disable-lock-screen dconf setting in order to hide the
default lock screen button in the system menu. But in Buster, this also
disables locking the screen via the org.gnome.ScreenSaver D-Bus service,
so we can't use this setting anymore.

This commit disables the disable-lock-screen dconf setting and hides the
default lock screen button via out gnome-shell extension instead.