Test suite: power off system under testing after scenario.
Until now we have relied on either one of the generated "snapshot
restore" steps or the "[Given] a computer" step to implicitly stop the
old VM when we move on to a new scenario. That meant the old VM was
still running during the new scenarios @Before@ hooks. If the new
scenario is tagged @check_tor_leaks that means we start its sniffer
while the old VM is still running, possibly sending packets that then
affect the new scenario. That would explain some myserious "Unexpected
connections were made" failures we have seen (refs: #11521).