Move boot instructions for PC and Mac to dedicated HTML inlines
I'll need that to be able to move the boot instructions for PC out of a
toggle and always displayed: for them to be hidden for Mac and visible
for PC, they have to be entirely inside an HTML tag. Until now it was
not needed only because we used a toogle who switch was itself only
displayed for PC.

Having dedicated inlines should make the code cleaner.