Really test that we're building anything but a stable release (refs: #12003)
This is a follow-up on commit:7174efed4c79c346d5f3dbd79a63086666866187,
that claimed to test whether we were building a stable release,
but merely checked whether $TAILS_CHANNEL == alpha. Both are equivalent
at the moment but as anonym pointed out, this assumption might become
wrong, e.g. if we ever add a "beta" channel.