Don't display the Home launcher on the desktop (refs: #16799)
Since the switch to the desktop-icons GNOME Shell extension, the nicer XDG-blah
name ("Home" in English, translated in many languages) is not used to label this
launcher anymore: instead, the name of the directory is displayed, in this case:
"amnesia", which makes no sense to our users.

Our other options to fix that are more costly and we've decided a while ago,
when I proposed to remove the desktop icons, to keep them until they were too
expensive to support. So this one goes: we have the Places menu already.