Test suite: only send TAB every second to get the syslinux kernel command-line (refs: #16820).
Our syslinux has a timeout of 5s so sending TAB every second should be enough
to guarantee we do open the kernel command line.

As anonym reported, "the spammer makes makes the splash show for significantly
longer: I've seen > 10x, so the boot splash never managed to appear, which is
worrying".