We should not write to the list.
List subscribers risk replying to the list instead of (or in addition to) to
tails@, which would violate "Unless you really want to start the key revocation
process, do not write to this mailing list".

I'd rather not rely only on a big fat warning to avoid that, while we
have a simple solution to lower the risk significantly.