Revert .fr and .de translations to the version from our canonical master branch.
It's not clear whether the .fr and .de translation teams consider
the work done on Weblate so far to be validated and better
than what we have in Git so let's stick to the translations
that went through our established process.

I'll let these translation teams know about this revert and then they
can choose what they do with it.