Add a VeraCrypt container with PIM for testing
For the automated VeraCrypt tests we create test containers on the fly.
But it's not easy to automatically create VeraCrypt containers with PIM in
Debian Stretch, so we add this one for use in the automated tests.

The password is 'test' and the PIM is 1. It contains a FAT filesystem
with a file 'GPL-3' copied from /usr/share/common-licenses/GPL-3 of a
Debian Stretch system.