Revert "Revert "Disable the uBlock logger sidebar.""
This reverts commit 114872f3c239fdbfef5cf07a92fbf5f3cdb62f6b,
i.e. it brings back the hack we had before we removed it in #16206.

Without this, the uBlock logger sidebar is displayed. I can reproduce this on
Debian sid with Firefox ESR but not with Firefox, which is consistent with
https://bugs.debian.org/909493#17 that reads "Please note this only works in
Firefox >= 62. The "open_at_install" option is not available in Firefox ESR".

I don't know why the upstream fix worked for us for a couple uBlock releases and
now stopped working, but whatever: once we ship Tor Browser 9.0 based on Firefox
68, we can remove this hack, so let's not bother figuring out why the hack is
needed for the next few months.