tailslib: Add tor_has_boostrapped()
I add this function to a new file systemd.sh instead of tor.sh, to make
it clear that this uses the systemd unit instead of communicating with
the Tor control socket (which is what tor.sh does).

This function should be used over tor_is_working() from tor.sh in
scripts which don't already communicate with the Tor control socket.