Add greeter files to POTFILES and refresh-translations
I added them the same way we added the unlock-veracrypt-volumes Python
and Glade files. This doesn't seem to be the ideal solution, because we
have to add files to POTFILES.skip which are not actually skipped. But
it should work for now.

This required renaming greeter.ui to greeter.ui.in.