Consistently suggest forking us on Salsa.
In the past we recommended forking us on gitlab.com, which was the only mirror
of our main Git repository (aka. tails.git) with a user-friendly "Fork" feature.

Since then, we have another such mirror — Salsa — that we actively use and keep
track of: for example, merge requests can be submitted there and the Foundations
Team will review them. This increases the chances that other teams start to do
the same there as well, until we've decided which GitLab instance will become
the canonical location for our code in the future.

This contrasts with gitlab.com, where we have disabled MRs on purpose.

So, let's point new contributors to a mirror where they can use the common "fork
and send a MR" workflow :)