Explicitly install gnome-shell to make the set of installed packages more deterministic.
This helps APT figure out that the dependency of the network-manager-gnome
package on "policykit-1-gnome | polkit-1-auth-agent" is already satisfied by
a Provides in the gnome-shell package. Otherwise, APT would install
policykit-1-gnome, that we don't need when running GNOME Shell.

See #16947 where we've spotted this problem "thanks" to a difference between the
set of packages installed when we build from a full (clone of the) Debian
archive vs. when we build from our partial, tagged APT repository.