Don't mask missing /usr/share/gdm/greeter/applications error case (refs: #12551).
Previously, "install -D" would create this directory if it does not exist.
Instead, let sed error out and abort the build if that's the case: this
directory is currently shipped by the gdm3 package; if that ever changes,
chances are that we need to move our gdm-shell-tails.desktop accordingly; better
notice that situation at build time, in a way that will point us towards the
root cause of the problem, instead of masking this change and then wondering why
our Greeter is not loaded correctly.