Remove very old known issue
Since then we added proper support for UEFI so this might only be useful
if very rare cases and only applicable by a very small fraction of our
target audience.

The current description of symptoms and possibly affected hardware is
extremely vague so I'm afraid of leading to many innocent people into a
dead end with this section.