Test suite: make the "[…] has loaded in the Tor Browser" step more robust (refs: #11592)
For some reason, sometimes the JavaScript that turns "Trying a testing version
of Tails" into "Tails" does not do its job. Waiting more does not help.
This seems to be a minor bug in Tor Browser or our website.

This is the primary cause of fragility of our test suite these days.

Given the benefit of getting more useful feedback from CI greatly outweighs
measuring how often that bug happens, and the scenarios that use this steps are
primarily about something else anyway, let's support both pre-JS and
post-JS titles.