Test suite: install Dogtail from Bullseye and run it with Python 3 (refs: #16976)
This will give us UTF-8 support.

Note the switch from raw bytes IO to StringIO: the equivalent of what we used to
do in Python 3 is io.BytesIO(), but that won't work out of the box: the code
we're running prints strings on stdout/stderr, not bytes, and Python 3 knows
the difference. So accordingly, remove decoding of the output, since we get
a string already.

Drop anonym's "showingOnly" patch that was merged upstream :)