Test suite: ensure we don't try clicking on a button before it's able to take our clicks (refs: #16004)
select_disk_image_dialog does not disappear immediately when one presses Enter
in there. And given it's displayed on top of restore_dialog, that contains
the button we want to click here, it prevents that click. showingOnly
does not help because the button we want to click *is* displayed:
it's just not clickable.