Test suite: drop useless code based on wrong assumptions (refs: #13470).
At least on Buster, as soon as gdm-password opens a session for the amnesia
user, a new session is registered with logind, then PostLogin is started; then,
once PostLogin has exited, LIVE_USER's gdm-x-session is started. So on Buster,
this step is almost a no-op: it will always succeed very quickly (if not on
first try) and only makes us start to wait for the desktop to be ready a tiny
bit later.