Test suite: make the "the support documentation page opens in Tor Browser" step more robust (refs: #15321)
I've seen this step fail while Tor Browser displays the correct page. That's
because the test was racy to me: after the "has loaded" step has passed, we're
immediately collecting a list of headings; but if Tor Browser is a bit slow, it
may not have exposed the entire page content on the a11y bus yet.

This problem is similar to the one fixed in commit
0d8a2b71774ee1e06788efcf8498a6d3acdd04cf, so let's apply the same solution here.