Make Spice file transfer reliable.
This fixes a race condition¹ that sometimes causes Spice file transfers to not
be available; I've seen this problem happen in the Journal saved by some test
suite runs.

Here, we apply the upstream fix² which was included in the spice-vdagent 0.19.0
release. Once we start building Tails based on Buster, this patch will fail to
apply, because the changes it brings are already applied, which will hint us
that we can remove it.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1623947
[2] https://lists.freedesktop.org/archives/spice-devel/2018-September/045511.html