Test suite: don't hard-code the list of locales that have a translated Unsafe Browser homepage (refs: #17004).
This list will change. For example, right now, no language has any
such translation, since the homepage was just fully rewritten.
When a language gets a translation, we can add a picture.
When a language loses its translation, we can remove the picture.
In both cases, we should not have to modify this step definition.