Test suite: make the "I set an administration password" step more robust.
On the devel branch, I've seen this failure mode: the first char of the admin
password is lost because we typed it before the corresponding text field was
displayed and focused. According to the debug log, we started typing the admin
password 170 ms after clicking TailsGreeterAdminPassword.png, so it's not
very surprising that occasionally we "win" this race against the Greeter.

Let's at least wait for the "Administration Password" dialog to be visible on
the screen before we interact with it.