Test suite: optimize performance of the "all notifications are disappeared" (refs: #17012)
The robustness fix in the previous commit makes this step quite slow in the most
common case — that is: a notification is displayed — because we'll have to wait
for Dogtail to give up waiting on the "No Notifications" label, which is nowhere
to be seen.

Let's optimize for the case where there is at least a notification, such as the
virtual machine warning, and look for the "Clear All" [notifications]
button first.

In passing, finish Dogtail'ifying this step by dropping its remaining
Sikuli call.