Test suite: provide guidance to fix problematic situation.
All my test runs with --keep-snapshots were failing because I had
a usb-install-tails-greeter snapshot in my tester VM, even though I had rebooted
it since that snapshot was created. But the corresponding file had been
deleted (it's in /tmp) so the snapshot could not be restored, and I was told "No
such (internal nor external) snapshot", which was rather unhelpful: it put me on
the wrong track (looking for something that was missing and why), while I should
have looked for an *existing* dangling snapshot to clean up.