Test suite: update 2 gedit images.
The previous version of these images is fuzzy on Buster. Most of the time,
Sikuli would retry, eventually find them on the screen, and save new versions in
sikuli_candidates, which is why we did not notice this problem and update the
reference images so far. But I've just seen one case when Sikuli could
not recognize GeditPaste.png, while is was plain to see on the screen,
so it's time to update these images.