Release process: fix Torrent publication.
When renaming the Torrent directories, I broke the part that was assuming that
the name of each such directory is the same as the name of the image
it contains.

And then, when uniting the Torrent folders, anonym broke the part that was
assuming we have a dedicated folder per image.

Closes: #16902