Release process: don't stay in a removed temporary directory.
Avoid leaving the RM's shell with a current directory set to a temporary
directory that's getting removed. Get back to the previous directory.

It's not as good as getting back to the actual starting point, but
that's still better than the current situation, without having to switch
to using subshells.