Test suite: don't restart Tor Browser between saving/opening (refs: #17056)
The "I close the Tor Browser" step is now broken when more than one tab is
opened, because Tor Browser 9.0 repaired the behavior controlled by the
browser.tabs.warnOnClose pref. In Tails 3.16 this pref was enabled but it was
kindy buggy: sometimes, no warning was displayed. With Tor Browser 9.0, I see
the "Quit and close tabs?" dialog pop up (as intended) more often, which totally
breaks this close+start sequence of steps, that relied on
browser.tabs.warnOnClose to be broken.

These extra steps were added in commit 47df1d7fbde1a66402781df46d5b6491534f0cba
with the following explanation:

  To make this scenario more robust we […] restart the browser between save/open
  in order to prevent Dogtail e.g. finding the old page that we just saved
  instead of the saved version we just opened.

My understanding is that the only case when Dogtail could possibly find the old
page, after we've opened a new tab where we're going to open the new page, is
when #12191 (Dogtail's showingOnly option is not always honored) happens.
So let's explicitly look for currently displayed frames, so that even when
that bug happens, Dogtail should not find the old page.