Install an Unsafe Browser red theme compatible with Firefox 68 (refs: #17055)
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/Themes#Documentation
says that "Lightweight themes have been deprecated", which explains why our old,
custom lightweight theme is ignored.

The only options to fix this are:

 - Create our own theme, get it automatically signed by Mozilla on AMO, download
   the resulting XPI, and somehow make it available at build time.

 - Use an existing theme with a GPL-3+ compliant license.

I've chosen the latter option: this theme is almost identical to the one we
had before.

The XPI is 7.7K so I don't think it's worth devising a more complicated
solution than just adding it to Git.