Refresh Thunderbird patch for Thunderbird 68 (refs: #16771)
This updates Make-use-of-non-SSL-Exchange-AutoDiscover-methods-op.patch
to apply cleanly on Thunderbird 68.

It seems like the "ssl_only_config_servers" pref is now named "sslOnly"
(see fd7df8a66cc10cc546bf7543c4905784125e34d4) so this commit renames that
pref in the patch.

Also, drop the hunk that patches the pref's comment. The comment in upstream is
different and I don't think it's worth our time to maintain the patch of the
comment (except if we plan to upstream it - do we?).