Don't accept an empty admin password in the Greeter
Previously, the user was able to accept an empty admin password in the
settings dialog. The password was not applied, instead it was disabled,
which is not what the user would expect. Fix this by not accepting an
empty password.

Also set the default widget status accordingly ("Add" button insensitive
when the dialog opens, password check mark hidden if both entries empty).