Fix bridge info not always shown when user selects bridge mode
When the user sets the network configuration to use bridges, we want to
inform them that they will have to configure the bridge later. But this
was only done when the popover was "built", which happens the first time
a network setting is configured. Now this information is always displayed
when the bridge mode is configured.