Test suite: bypass Tor Browser GUI issues and open URLs via the command line (refs: #17056)
On our slower isotesters, pasting the URL does not automatically open the
Visit/Search menu, and the Enter key press is lost. I've also tried to use
"Paste & Go" and to type the URL instead of pasting in; none of those helped.

So let's have something that works, even if arguably it's not in the spirit of
BDD. OTOH, it exercises opening URLs in the same way as they would if the user
clicked on a link from another application, which is useful too.