Disable/break Tor Browser self-updates (Closes: #17114)
app.update.enabled is not supported anymore:
https://bugzilla.mozilla.org/show_bug.cgi?id=1420514

So for example, without this commit, an image built with Tor Browser 9.0a6
would complain that it's outdated.

Debian's Firefox passes --disable-updater to the build system,
but we can't do this as we use upstream Tor Browser builds.

So instead, let's do this:

 - Point app.update.url to a non-existent place, so that new updates
   cannot be found, even if Tor Browser tries to check for updates
   despite our other attempts to disable that.

 - Enable app.update.disabledForTesting, which is what Mozilla replaced
   app.update.enabled with in their test suite.

 - Disable app.update.doorhanger, so any update found despite everything else
   won't be advertised in the UI.

 - Disable app.update.auto, so that Tor Browser does not install automatically
   any update it may find despite everything else.

Among those, any of the 2 first prefs should be enough, but let's play it safe.

The alternative would be to use policies.json:

  https://support.mozilla.org/en-US/kb/customizing-firefox-using-policiesjson

… but let's try to adding another moving piece to our Tor Browser puzzle, if we
can avoid it. Besides, it's already been broken in Tor Browser:

  https://trac.torproject.org/projects/tor/ticket/29445
  https://gitweb.torproject.org/user/brade/tor-browser.git/commit/?h=bug29445-01&id=7ce43156780e3b91f35481cce1eb4afed56a4b0f

… and I'm warry it could be broken again without anyone noticing
early enough.