Browsers: disable the Quantum Bar.
Disabling it seems prevent a nasty bug observed on primarily
Jenkins (but also elsewhere) where the URL bar doesn't respond to
Enter and the suggestions pop-up doesn't show up.

Modifying this pref only changes which tech is used under the hood for
displaying the URL bar [0] so nothing will be different from users.

[0] https://www.ghacks.net/2019/06/27/this-is-firefoxs-new-quantumbar-address-bar/

Refs: #17121