Test suite: do a clean reboot in the Persistent browser bookmarks scenario (refs: #17125)
commit ec8c3541334e4e822a54042326318db300352b51 repaired what it meant to, but
it make the next steps of this scenario fail: our "I cold reboot the computer"
step runs "I power off the computer", which runs $vm.power_off, which runs
@domain.destroy ⇒ Tails has no time to sync files to disk and the bookmark we
meant to save to persistence is nowhere to be found after rebooting.

So let's do a clean reboot in "I cold reboot the computer", which is not
used anywhere else anyway.