Test suite: only partially fill memory for userspace processes.
Filling all memory often leads to the VM crashing. So why fill all
memory? Back when Tails was 32-bit (or mixed with 64-bit) there were
some of uncertainty around the 4 GiB limit, which motivated us to fill
a larger region than 4 GiB, so without much thought we might have just
gone ahead and filled all of it. But, AFAICT, if we went with the
approach I describe above and allocated >4 GiB then the test would
still have done what we wanted.

In today's fully 64-bit world we have no reason for allocating >4 GiB,
so we could go with 128 MiB like the other tests arbitrarily (?) do.

Will-fix: #17104