RM'ing: call for manual testers for Tails vN+1 immediately after releasing Tails vN.
Rationale: we're operating with a pool of manual testers that's dangerously
small, in particular considering we'll switch to a 4-weeks release cadence soon.
One potential explanation for this is that we often announce late in the release
cycle the date when we'll need testers. Let's try to increase the chances
testers can book time for this, by telling them about it earlier.

This was discussed by the RM team in the "RMs doing more and more manual tests
[Was: 3.16 post-releasem]" thread. The same idea was suggested independently by
someone else on tails@.