Release schedule: remove dead link.
I've tried to find this post and the best I could find is
https://lists.autistici.org/message/20120428.185212.f0ff0ca2.en.html,
where I advocated in favour of a time-based release schedule,
for a variety of reasons.

Since then, we've consistently aligned our release schedule with Tor Browser's
(and thus, transitively, with Firefox's) in order to let our users down with
known security vulnerabilities in the most important app we ship: the
web browser.

So at this point, the other reasons why one may prefer a time-based release
schedule don't matter much anymore.