Test suite: drop the "Unsafe Browser has no proxy configured" step (refs: #17157)
Our firewall blocks access from the clearnet user to the loopback interface,
so proxies are not supposed to work. We validate this already with
a combination of two tests:

 - "Scenario: The Unsafe Browser can load a web page" would fail
   if a proxy was configured and the firewall is set up as intended.

 - "Scenario: The Unsafe Browser cannot be configured to use Tor and other local
   proxies" would fail if the firewall was not set up correctly.

So let's remove this extra test, which is hard to update since the UI it's
currently using was removed from Tor Browser 9.0a8's code base.