Test suite: update Unsafe Browser proxy test for Tor Browser 9.0a8 (refs: #17157)
The UI this test used to rely on was removed from Tor Browser 9.0a8's
code base, so let's re-implement it via prefs.js.

And in passing, only test one Socksport (chosen randomly among those configured
in torrc): this should be sufficient to verify that our firewall does not allow
the clearnet user to connect to services listening on the loopback interface,
which is what this scenario is essentially about.