Don't implicitly rely on the fact this glob is not expanded by the current shell.
If the shell can expand this glob, then it will. If it can't, then
it will pass it as-is to 7z, which is what we want.

In practice, the chances this glob will ever expand is minuscule, but
developers (at least myself) would have to wonder about this every time they
work on this piece of code, which is a little bit distracting.

Fixup against 09fcbc152a021485773c3310180c4b6fedef6813.