Test suite: ensure we don't break tests by opening the Applications menu in post_vm_start_hook (refs: #17164)
In 25% of the last 16 test suite runs on the testing branch, the "Starting
a second instance of the Unsafe Browser results in an error message being shown"
scenario failed because the Applications menu had been opened by
post_vm_start_hook, which breaks opening the Activities Overview: pressing the
SUPER key has no effect when the Applications menu is opened.

To workaround this problem, this commit uses the same trick that's already used
in the "all notifications have disappeared" step: press ESC and wait 1 second.

I admit I don't understand why Sikuli clicking on (1023,384) can possibly open
the Applications menu. It could be a bug in GNOME Shell, it could be a bug in
Sikuli, it could be a race condition. Given we'll ditch Sikuli soonish, and we
have a good workaround readily available, I won't invest more time into
investigating this now.