Vagrant: remove obsolete hack
This hack was originally added to fix #12530, that is, to ensure that services
won't start when we install packages in postinstall.sh. It's not needed anymore
since we now have /proc in the chroot, which enables detection of chroot by
various pieces of code, including the one that would otherwise start services.