build-tails: wait for NTP to be disabled before setting the desired date (refs: #16868)
"timedatectl set-ntp false" merely queues up a job: it does not wait until NTP
is actually disabled. So far we did not notice that this was racing against us
setting a custom date, but systemd 241 it becomes obvious: systemd does not
allow other operations while one did not finish anymore.

Disabling NTP services can take a bit of time, so with a Buster Vagrant box,
when $TAILS_DATE_OFFSET is set (which we do for our
reproducibly_build_Tails_ISO_* Jenkins jobs), the build would fail like this:

 + as_root_do timedatectl set-ntp false
 + sudo http_proxy=http://192.168.122.10:3142 APT_SNAPSHOTS_SERIALS={"torproject":"2019100904","debian":"2019100904","debian-security":"2019101901"} TAILS_MERGE_BASE_BRANCH=1 GIT_COMMIT=844ef7072f5856be50635aba45d6ebd820e70205 GIT_REF=feature/16868-buster-vagrant-box+force-all-tests BASE_BRANCH_GIT_COMMIT=b2559c50ba6c3f839da89d475964160889f8fbad timedatectl set-ntp false
 + date --utc --date=+8 days +%F %T
 Setting system time to 2019-10-27 09:52:54
 + DESIRED_DATE=2019-10-27 09:52:54
 + echo Setting system time to 2019-10-27 09:52:54
 + as_root_do timedatectl set-time 2019-10-27 09:52:54
 + sudo http_proxy=http://192.168.122.10:3142 APT_SNAPSHOTS_SERIALS={"torproject":"2019100904","debian":"2019100904","debian-security":"2019101901"} TAILS_MERGE_BASE_BRANCH=1 GIT_COMMIT=844ef7072f5856be50635aba45d6ebd820e70205 GIT_REF=feature/16868-buster-vagrant-box+force-all-tests BASE_BRANCH_GIT_COMMIT=b2559c50ba6c3f839da89d475964160889f8fbad timedatectl set-time 2019-10-27 09:52:54
 Failed to set time: Previous request is not finished, refusing.

So let's wait for the NTP service to be disabled before using timedatectl to
configure stuff again.