Test suite: make the "GnuPG's dirmngr uses the configured keyserver" step actually test what it is meant to
On #12371 we discussed how to query the keyserver that's actually used
by dirmngr. In the end we seem to have found a solution that should work.
But the commit that was pushed to fix that ticket included what looks
like leftovers from previous attempts, which made the code we had decided
to use a no-op, masked by code that does not test what we meant to test:
f31e2b823607ca6138e1437d5b02e47969cb4e03.

So let's drop this leftover code in favour of the one we actually wanted
to use.