Test suite: ensure dirmngr picks up the changes we make to its configuration
This is not really specific to #12689: depending on when dirmngr is first
used (be it by our test suite or some other Tails background process), and thus
auto-started via dirmngr.socket, it may have loaded dirmngr.conf already.
If we lose this race, then the changes we apply to this file in this step
won't be taken into account by dirmngr, which can result in hard to debug
test suite failures.

So instead, let's ensure our (so far implicit) assumption is correct,
that is: dirmngr uses the updated version of dirmngr.conf.