Test suite: switch backend keyservers (refs: #12689, #14770)
First, we do have to do _something_ here, as long as we use an Onion service as
our default keyserver: our Chutney is not able to connect to Onion
services (#12210).

1. dirmngr: use keys.openpgp.org's Onion service directly
=========================================================

That is, without going through a Chutney-based Onion service.

When dirmngr connects to the Onion service run by Chutney, the isotester
redirects the connection to keys.openpgp.org:11371; so far, so good. But then,
keys.openpgp.org redirects us to https://keys.openpgp.org, and the key retrieval
fails for some reason I don't fully understand:

   dirmngr[10130]: connection from process 10145 (1000:1000)
   dirmngr[10130]: DBG: chan_5 <- GETINFO version
   dirmngr[10130]: DBG: chan_5 -> D 2.2.12
   dirmngr[10130]: DBG: chan_5 -> OK
   dirmngr[10130]: DBG: chan_5 <- KS_GET -- 0xC4BC2DDB38CCE96485EBE9C2F20691179038E5C6
   dirmngr[10130]: DBG: gnutls:L3: ASSERT: ../../../lib/x509/common.c[_gnutls_x509_get_raw_field2]:1570
   dirmngr[10130]: DBG: gnutls:L3: ASSERT: ../../../lib/x509/x509.c[gnutls_x509_crt_get_subject_unique_id]:3897
   dirmngr[10130]: DBG: gnutls:L3: ASSERT: ../../../lib/x509/x509.c[gnutls_x509_crt_get_issuer_unique_id]:3947
   dirmngr[10130]: DBG: gnutls:L3: ASSERT: ../../../lib/x509/dn.c[_gnutls_x509_compare_raw_dn]:990
   dirmngr[10130]: number of system provided CAs: 128
   dirmngr[10130]: DBG: gnutls:L5: REC[0x78432c2d1e10]: Allocating epoch #0
   dirmngr[10130]: DBG: gnutls:L2: added 6 protocols, 29 ciphersuites, 18 sig algos and 9 groups into priority list
   dirmngr[10130]: URL 'http://nl5vtjfpfz2llza7.onion:5858/pks/lookup?op=get&options=mr&search=0xC4BC2DDB38CCE96485EBE9C2F20691179038E5C6' redirected to 'https://keys.openpgp.org/pks/lookup?op=get&options=mr&search=0xC4BC2DDB38CCE96485EBE9C2F20691179038E5C6' (301)
   dirmngr[10130]: DBG: gnutls:L5: REC[0x78432c2d1e10]: Start of epoch cleanup
   dirmngr[10130]: DBG: gnutls:L5: REC[0x78432c2d1e10]: End of epoch cleanup
   dirmngr[10130]: DBG: gnutls:L5: REC[0x78432c2d1e10]: Epoch #0 freed
   dirmngr[10130]: command 'KS_GET' failed: Forbidden <Unspecified source>
   dirmngr[10130]: DBG: chan_5 -> ERR 251 Forbidden <Unspecified source>
   dirmngr[10130]: DBG: chan_5 <- BYE
   dirmngr[10130]: DBG: chan_5 -> OK closing connection

dirmngr has a rather complex history wrt. TLS certificate validation vs.
HTTP redirections vs. DNS resolution, so I'm not surprised that this fails.

The code I'm disabling here was introduced for #12211. As I understand it, the
main goal there was to adjust our OpenPGP test cases to the fact we had switched
Tails to a .onion keyserver. It's not 100% clear to me why we added the layer of
indirection I'm removing here, instead of doing something similar to this
commit: I see no explanation on the ticket nor in the corresponding
commit messages.

So let's drop the extra complexity of going through a proxy Onion service on
Chutney, and instead reconfigure dirmngr to directly connect to
keys.openpgp.org. AFAICT, given previously we were not testing the dirmngr
configuration we're shipping to our users either, the only test coverage we lose
here is: testing that dirmngr can connect to a Onion service.

2. Seahorse: use a keyserver that meets our (so far implicit) requirements
==========================================================================

At least one member of pool.sks-keyservers.net does not satisfy the implicit
assumption that the code added for #12211 relied upon. This commit fixes that
by always using a backend, upstream keyserver that satisfies our assumptions.

Other than that, we stick to the previous setup here. This seems to be the only
vaguely viable option given we can't configure Seahorse to use keys.openpgp.org:
that keyserver redirects to HTTPS, which Seahorse does not support.

I'm (re-)tagging the Seahorse scenarios @fragile: in the end, it's not clear
whether this branch will improve things for those scenarios.